+++ e2e-tests/tests/angular-17.test.ts
@@ -110,6 +110,13 @@ async function runWizardOnAngularProject(
   await wizardInstance.sendStdinAndWaitForOutput(
     // select "Yes" for replay
     [KEYS.ENTER],
+    // "Do you want to enable Logs", sometimes doesn't work as `Logs` can be printed in bold.
+    'to send your application logs to Sentry?',
+  );
+
+  await wizardInstance.sendStdinAndWaitForOutput(
+    // select "Yes" for logs
+    [KEYS.ENTER],
     'Where are your build artifacts located?',
     {
       timeout: 240_000, // installing Sentry CLI can take a while
@@ -216,6 +223,7 @@ function checkAngularProject(
       'tracesSampleRate: 1',
       'replaysSessionSampleRate: 0.1',
       'replaysOnErrorSampleRate: 1',
+      'enableLogs: true',
     ]);
   });